// ui-service/src/app/components/reporting/model/column-model.ts
export enum ReportColumn {
  ID = 'id',
  CODE = 'code',
  TITLE = 'title',
  DESCRIPTION = 'description',
  URL_SUFFIX = 'urlSuffix',
  SERVICE_NAME = 'serviceName',
  ENABLED = 'enabled',
  DEFAULT_PARAMETERS = 'defaultParameters',
  LAST_EXECUTION_START_TIME = 'lastExecutionStartTime',
  LAST_EXECUTION_FINISH_TIME = 'lastExecutionFinishTime',
  LAST_EXECUTION_RESULT_DATA = 'lastExecutionResultData',
  LAST_EXECUTION_PARAMETERS = 'lastExecutionParameters',

  EXECUTE = 'execute',
}